+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.60](https://github.com/MarcoIeni/release-plz/compare/release-plz-v0.3.59...release-plz-v0.3.60) - 2024-04-14
+
+### Fixed
+- correctly alias `project-manifest` ([#1396](https://github.com/MarcoIeni/release-plz/pull/1396))
+
+## [0.3.59](https://github.com/MarcoIeni/release-plz/compare/release-plz-v0.3.58...release-plz-v0.3.59) - 2024-04-14
+
+### Other
+- Change the flag --project-manifest to --manifest-path [#1333](https://github.com/MarcoIeni/release-plz/pull/1333) ([#1390](https://github.com/MarcoIeni/release-plz/pull/1390)) ([#1395](https://github.com/MarcoIeni/release-plz/pull/1395))
+
+## [0.3.58](https://github.com/MarcoIeni/release-plz/compare/release-plz-v0.3.57...release-plz-v0.3.58) - 2024-04-08
+
+### Added
+- add json output ([#1386](https://github.com/MarcoIeni/release-plz/pull/1386)) ([#1368](https://github.com/MarcoIeni/release-plz/pull/1368))
+
+### Other
+- document `get_registry_packages` function ([#1381](https://github.com/MarcoIeni/release-plz/pull/1381))
+
 ## [0.3.57](https://github.com/MarcoIeni/release-plz/compare/release-plz-v0.3.56...release-plz-v0.3.57) - 2024-04-02
 
 ### Fixed